courses page ui
