+++ app/src/main/java/com/kyrgyzbilim/base/InjectorObject.kt
@@ -1,0 +1,13 @@
+package com.kyrgyzbilim.base
+
+import com.kyrgyzbilim.data.repository.CourseRepository
+import com.kyrgyzbilim.data.repository.CourseRepositoryImpl
+import com.kyrgyzbilim.ui.courses.CourseViewModel
+
+object InjectorObject {
+
+//    val courseRepository: CourseRepository = CourseRepositoryImpl(service)
+//
+//    fun getMainViewModelFactory() = CourseViewModel(courseRepository)
+
+}
