+++ app/src/main/java/com/kyrgyzbilim/data/Course.kt
@@ -1,0 +1,14 @@
+package com.kyrgyzbilim.data
+
+class Course(
+        val id: Int,
+        val levelTitle: String,
+        val description: String,
+        val progress : Int
+
+        ) {
+
+        override fun equals(other: Any?): Boolean {
+                return super.equals(other)
+        }
+}
