+++ app/src/main/java/com/kyrgyzbilim/data/repository/CourseRepository.kt
@@ -1,0 +1,9 @@
+package com.kyrgyzbilim.data.repository
+
+import com.kyrgyzbilim.data.Course
+
+class CourseRepository {
+
+//    suspend fun getCourses(courseId: Int): ApiResult<Course>
+
+}
