+++ app/src/main/java/com/kyrgyzbilim/data/repository/CourseRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.kyrgyzbilim.data.repository
+
+class CourseRepositoryImpl {
+}
