+++ app/src/main/java/com/kyrgyzbilim/ui/authorization/RegistrationFragment.kt
@@ -1,0 +1,25 @@
+package com.kyrgyzbilim.ui.authorization
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.kyrgyzbilim.R
+
+
+class RegistrationFragment : Fragment() {
+    // TODO: Rename and change types of parameters
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_registration, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
