+++ app/src/main/java/com/kyrgyzbilim/ui/courses/CourseViewModel.kt
@@ -1,0 +1,4 @@
+package com.kyrgyzbilim.ui.courses
+
+class CourseViewModel {
+}
