+++ app/src/main/java/com/kyrgyzbilim/ui/courses/CourseViewModelFactory.kt
@@ -1,0 +1,4 @@
+package com.kyrgyzbilim.ui.courses
+
+class CourseViewModelFactory {
+}
