--- app/src/main/java/com/taverapp/kyrgyzbilim/data/Course.kt
@@ -1,10 +1,0 @@
-package com.taverapp.kyrgyzbilim.data
-
-class Course(
-        val courseName: String,
-        val description: String,
-        val progress : Int
-
-        ) {
-
-}
