--- app/src/main/java/com/taverapp/kyrgyzbilim/ui/courses/CoursesFragment.kt
@@ -1,30 +1,0 @@
-package com.taverapp.kyrgyzbilim.ui.courses
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.core.content.ContextCompat
-import com.taverapp.kyrgyzbilim.R
-import kotlinx.android.synthetic.main.item_courses.*
-
-
-class CoursesFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_courses, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-
-
-    }
-}
