--- app/src/main/java/com/taverapp/kyrgyzbilim/ui/themesSection/grammar/GrammarFragment.kt
@@ -1,25 +1,0 @@
-package com.taverapp.kyrgyzbilim.ui.themesSection.grammar
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.taverapp.kyrgyzbilim.R
-
-
-class GrammarFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_grammar, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
