+++ app/src/main/res/layout/item_courses.xml
@@ -6,16 +6,14 @@
     android:layout_height="wrap_content">
 
     <androidx.cardview.widget.CardView
+        android:layout_marginTop="4dp"
         android:layout_width="370dp"
         android:layout_height="175dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
+        android:layout_marginStart="18dp"
+        android:layout_marginEnd="18dp"
         app:cardCornerRadius="8dp"
-        app:cardElevation="15dp"
-        android:layout_marginBottom="12dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:cardElevation="10dp"
+        android:layout_marginBottom="12dp">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -29,7 +27,7 @@
                 android:orientation="horizontal">
 
                 <TextView
-                    android:id="@+id/level"
+                    android:id="@+id/level_title"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="16dp"
@@ -87,6 +85,7 @@
                     android:id="@+id/percent_progress"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
+                    android:textSize="12sp"
                     android:gravity="center_vertical"
                     android:text="50%"
                     android:textColor="@color/progressBar"
